ShopStrategy class
